requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
apscheduler==3.10.4
caldav==1.3.9
icalendar==6.0.1
httpx==0.27.2
PyJWT==2.9.0
pytz==2024.1
twilio==9.3.6
